--- src/main/java/com/mobble/mobbleserver/account/auth/principal/AuthMember.java
@@ -1,4 +1,0 @@
-package com.mobble.mobbleserver.account.auth.principal;
-
-public record AuthMember(Long memberId) {
-}
