+++ src/main/java/com/mobble/mobbleserver/infrastructure/jwt/principal/AuthMember.java
@@ -1,0 +1,4 @@
+package com.mobble.mobbleserver.infrastructure.jwt.principal;
+
+public record AuthMember(Long memberId) {
+}
